Time: 8 ms (6.98%) | Memory: 8 MB (5.54%) - LeetSync

+++ 535-encode-and-decode-tinyurl/encode-and-decode-tinyurl.cpp
@@ -16,8 +16,8 @@ public:
     }
 
     private:
-    uint64_t hash_function( string& str) {
-      uint64_t hash = 9239;
+    long long int hash_function( string& str) {
+      long long int hash = 9239;
 
         for(int i = 0; i < str.size(); ++i) {
            uint64_t value = str[i];
